+++ cache/lru/core/lru_interface.go
